+++ src/components/LiquidityChartRangeInput/hooks.ts
@@ -1,6 +1,6 @@
 import { Currency } from '@uniswap/sdk-core'
 import { FeeAmount } from '@uniswap/v3-sdk'
-import { TickProcessed, usePoolActiveLiquidity } from 'hooks/usePoolActiveLiquidity'
+import { TickProcessed, usePoolActiveLiquidity } from 'hooks/usePoolTickData'
 import { useCallback, useMemo } from 'react'
 
 import { ChartEntry } from './types'
